requirements.txt
PyMySQL3==0.5
